#  common/trinary: trits and bytes conversions optimizations