--- a/common/trinary/trit_ptrit.h
+++ b/common/trinary/trit_ptrit.h
@@ -5,10 +5,6 @@
  * Refer to the LICENSE file for licensing information
  */
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 #ifndef __COMMON_TRINARY_TRIT_PTRIT_H_
 #define __COMMON_TRINARY_TRIT_PTRIT_H_
 
@@ -16,14 +12,19 @@ extern "C" {
 #include "common/trinary/ptrit_incr.h"
 #include "common/trinary/trits.h"
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 void trits_to_ptrits(trit_t const *const trits, ptrit_t *const ptrits,
                      size_t const index, size_t const length);
 void trits_to_ptrits_fill(trit_t const *const trits, ptrit_t *const ptrits,
                           size_t const length);
 void ptrits_to_trits(ptrit_t const *const ptrits, trit_t *const trits,
-                     size_t const index, size_t length);
+                     size_t const index, size_t const length);
 
-#endif
 #ifdef __cplusplus
 }
 #endif
+
+#endif  // __COMMON_TRINARY_TRIT_PTRIT_H_