--- a/common/trinary/trit_tryte.c
+++ b/common/trinary/trit_tryte.c
@@ -9,14 +9,13 @@
 
 #include "common/trinary/trit_tryte.h"
 
-static const trit_t
-    TRYTE_TO_TRITS_MAPPINGS[TRYTE_SPACE][NUMBER_OF_TRITS_IN_A_TRYTE] = {
-        {0, 0, 0},   {1, 0, 0},   {-1, 1, 0},  {0, 1, 0},   {1, 1, 0},
-        {-1, -1, 1}, {0, -1, 1},  {1, -1, 1},  {-1, 0, 1},  {0, 0, 1},
-        {1, 0, 1},   {-1, 1, 1},  {0, 1, 1},   {1, 1, 1},   {-1, -1, -1},
-        {0, -1, -1}, {1, -1, -1}, {-1, 0, -1}, {0, 0, -1},  {1, 0, -1},
-        {-1, 1, -1}, {0, 1, -1},  {1, 1, -1},  {-1, -1, 0}, {0, -1, 0},
-        {1, -1, 0},  {-1, 0, 0}};
+static const trit_t TRYTES_TRITS_LUT[TRYTE_SPACE][NUMBER_OF_TRITS_IN_A_TRYTE] =
+    {{0, 0, 0},   {1, 0, 0},   {-1, 1, 0},  {0, 1, 0},   {1, 1, 0},
+     {-1, -1, 1}, {0, -1, 1},  {1, -1, 1},  {-1, 0, 1},  {0, 0, 1},
+     {1, 0, 1},   {-1, 1, 1},  {0, 1, 1},   {1, 1, 1},   {-1, -1, -1},
+     {0, -1, -1}, {1, -1, -1}, {-1, 0, -1}, {0, 0, -1},  {1, 0, -1},
+     {-1, 1, -1}, {0, 1, -1},  {1, 1, -1},  {-1, -1, 0}, {0, -1, 0},
+     {1, -1, 0},  {-1, 0, 0}};
 
 trit_t get_trit_at(tryte_t const *const trytes, size_t const length,
                    size_t const index) {
@@ -25,9 +24,9 @@ trit_t get_trit_at(tryte_t const *const trytes, size_t const length,
     return 0;
   }
   tryte_t tryte = trytes[tindex];
-  size_t cindex = tryte == '9' ? 0 : tryte - '@';
+  size_t cindex = INDEX_OF_TRYTE(tryte);
   tindex = index % 3U;
-  return TRYTE_TO_TRITS_MAPPINGS[cindex][tindex];
+  return TRYTES_TRITS_LUT[cindex][tindex];
 }
 
 uint8_t set_trit_at(tryte_t *const trytes, size_t const length,
@@ -37,13 +36,13 @@ uint8_t set_trit_at(tryte_t *const trytes, size_t const length,
     return 0;
   }
   tryte_t tryte = trytes[tindex];
-  size_t cindex = tryte == '9' ? 0 : tryte - '@';
+  size_t cindex = INDEX_OF_TRYTE(tryte);
   trit_t trits[3];
   if (cindex > 26) {
     // uninitialized
-    memcpy(trits, TRYTE_TO_TRITS_MAPPINGS[0], NUMBER_OF_TRITS_IN_A_TRYTE);
+    memcpy(trits, TRYTES_TRITS_LUT[0], NUMBER_OF_TRITS_IN_A_TRYTE);
   } else {
-    memcpy(trits, TRYTE_TO_TRITS_MAPPINGS[cindex], NUMBER_OF_TRITS_IN_A_TRYTE);
+    memcpy(trits, TRYTES_TRITS_LUT[cindex], NUMBER_OF_TRITS_IN_A_TRYTE);
   }
   size_t uindex = index % 3U;
   trits[uindex] = trit;
@@ -59,13 +58,13 @@ void trits_to_trytes(trit_t const *const trits, tryte_t *const trytes,
                      size_t const length) {
   int k = 0;
 
-  for (size_t i = 0, j = 0; i < length; i += 3, j++) {
+  for (size_t i = 0, j = 0; i < length; i += RADIX, j++) {
     k = 0;
     for (size_t l = length - i < NUMBER_OF_TRITS_IN_A_TRYTE
                         ? length - i
                         : NUMBER_OF_TRITS_IN_A_TRYTE;
          l-- > 0;) {
-      k *= 3;
+      k *= RADIX;
       k += trits[i + l];
     }
     if (k < 0) {
@@ -81,9 +80,8 @@ void trytes_to_trits(tryte_t const *const trytes, trit_t *const trits,
     return;
   }
 
-  for (size_t i = 0, j = 0; i < length; i++, j += 3) {
-    memcpy(trits + j,
-           TRYTE_TO_TRITS_MAPPINGS[trytes[i] == '9' ? 0 : trytes[i] - '@'],
+  for (size_t i = 0, j = 0; i < length; i++, j += RADIX) {
+    memcpy(trits + j, TRYTES_TRITS_LUT[INDEX_OF_TRYTE(trytes[i])],
            NUMBER_OF_TRITS_IN_A_TRYTE);
   }
 }