--- a/common/trinary/tryte_long.h
+++ b/common/trinary/tryte_long.h
@@ -5,20 +5,21 @@
  * Refer to the LICENSE file for licensing information
  */
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 #ifndef __COMMON_TRINARY_TRYTE_LONG_H_
 #define __COMMON_TRINARY_TRYTE_LONG_H_
 
 #include "common/trinary/tryte.h"
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 size_t min_trytes(int64_t const value);
-int64_t trytes_to_long(tryte_t const *const trytes, size_t const i);
+int64_t trytes_to_long(tryte_t const *const trytes, size_t const length);
 size_t long_to_trytes(int64_t const value, tryte_t *const trytes);
 
-#endif  // __COMMON_TRINARY_TRYTE_LONG_H_
 #ifdef __cplusplus
 }
 #endif
+
+#endif  // __COMMON_TRINARY_TRYTE_LONG_H_