… common/trinary/tests/test_trit_byte.c
… common/trinary/tryte_long.c