- Various nits
- Optimizes trits and bytes conversions

100000000 iterations
trits_to_bytes  before  8755ms
trits_to_bytes  after   5392ms (-38.412%)

100000000 iterations
bytes_to_trits  before  11844ms
bytes_to_trits  after   5516ms (-53.428%)

A lookup table has been introduced for trits and bytes conversions. Since its size can be quite a bit for very small devices, it can be disabled with a `#define` and it defaults to recomputing values at runtime.

# Test Plan:
CI